Split trainer into its own class, which is used by the classifier training script
