+++ src/utils/common_utils.py
@@ -1,0 +1,6 @@
+from datetime import datetime
+
+_all__ = ["log"]
+
+def log(message):
+  print(f'{datetime.now().isoformat()} {message}')
